People can't see button unless he is group owner

+++ resources/views/groups/index.blade.php
@@ -22,8 +22,10 @@
 	          <td>{{ $group->description }}</td>
 	          <td>{{ $group->user->name }}</td>
 	          <td>
-		          <a href="{{ route('groups.edit', compact('group')) }}" class="btn btn-sm btn-default">編輯</a>
-		          <a href="{{ route('groups.destroy', compact('group')) }}" class="btn btn-sm btn-default" data-method="delete" data-confirm="Are you sure?" data-token="{{csrf_token()}}">刪除</a>
+	          	@can('update', $group)
+			          <a href="{{ route('groups.edit', compact('group')) }}" class="btn btn-sm btn-default">編輯</a>
+			          <a href="{{ route('groups.destroy', compact('group')) }}" class="btn btn-sm btn-default" data-method="delete" data-confirm="Are you sure?" data-token="{{csrf_token()}}">刪除</a>
+		          @endcan
 	          </td>
 	        </tr>
 	      @endforeach
